Fixed reading of input command
The reading of the command failed to read the line return, resulting in a double printing of the input menu.  Switching to using the "readline" command eliminated this as well as provides some extract protection against extraneous data entry.

The function "flushInput" was supposed to clear any extra data, but it did not work in my testing.  The text entered from the serial monitor had not arrived at the software serial buffer, so the loop starting "while(ss.available())" was never entered.  Then data entered later, then was read at the next loop as a new (unrecognized) command, therefore cause an entry error followed by another trip around the loop which reprinted the menu (resulting in the double menu print).

The "flushInput" function was no longer needed so it was removed.